refactoring led emitter to just be a driver. Renamed to led_driver. Led emitter logic is moved to dt-core
