+++ packages/duckiebot_interface/launch/all_drivers.launch
@@ -28,9 +28,9 @@
         </group>
     </group>
 
-    <!-- LED Emitter: reads ROS messages and commands the LEDs -->
+    <!-- LED Driver: reads ROS messages and commands the LEDs -->
     <group if="$(eval arg('robot_type') in ['duckiebot', 'traffic_light'])">
-        <include file="$(find led_emitter)/launch/led_emitter_node.launch">
+        <include file="$(find led_driver)/launch/led_driver_node.launch">
             <arg name="veh" value="$(arg veh)"/>
             <arg name="robot_type" value="$(arg robot_type)"/>
         </include>
