+++ packages/led_driver/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 2.8.3)
-project(led_emitter)
+project(led_driver)
 
 
 find_package(catkin REQUIRED COMPONENTS
