+++ packages/led_driver/launch/led_driver_node.launch
@@ -1,12 +1,12 @@
 <launch>
-	<arg name="node_name" value="led_emitter_node"/>
+	<arg name="node_name" value="led_driver_node"/>
 	<arg name="veh" doc="Name of vehicle. ex: megaman"/>
   <arg name="param_file_name" default="LED_protocol" doc="Specify a param file. ex:megaman"/>
 	<arg name="robot_type" default="duckiebot" doc="Type of robot"/>
 
 	<group ns="$(arg veh)">
-	    <node pkg="led_emitter" type="led_emitter_node.py"  name="$(arg node_name)" output="screen" required="true">
-			<rosparam command="load" file="$(find led_emitter)/config/$(arg node_name)/$(arg param_file_name).yaml"/>
+	    <node pkg="led_emitter" type="led_driver_node.py"  name="$(arg node_name)" output="screen" required="true">
+			<rosparam command="load" file="$(find led_driver)/config/$(arg node_name)/$(arg param_file_name).yaml"/>
 			<param name="robot_type" value="$(arg robot_type)"/>
 		</node>
 	</group>
