+++ packages/led_driver/setup.py
@@ -5,8 +5,8 @@ from catkin_pkg.python_setup import generate_distutils_setup
 
 # fetch values from package.xml
 setup_args = generate_distutils_setup(
-    packages=['rgb_led'],
-    package_dir={'': 'include'},
+    packages=["rgb_led"],
+    package_dir={"": "include"},
 )
 
 setup(**setup_args)
